Put grid to Roomdata script
