+++ Assets/MenuUi/Scripts/SoulHome/RoomData.cs
@@ -1,12 +1,14 @@
 using System;
 using System.Collections;
 using System.Collections.Generic;
-using UnityEngine;
-using MenuUI.Scripts.SoulHome;
-using Debug = Prg.Debug;
+using System.Runtime.CompilerServices;
 using Altzone.Scripts.Model.Poco.Game;
 using Altzone.Scripts.ReferenceSheets;
+using MenuUI.Scripts.SoulHome;
+using UnityEditor;
+using UnityEngine;
 using UnityEngine.Rendering;
+using Debug = Prg.Debug;
 
 namespace MenuUI.Scripts.SoulHome
 {
@@ -29,6 +31,8 @@ namespace MenuUI.Scripts.SoulHome
         [SerializeField]
         private int _slotColumns = 8;
         private int _slotHeight = 6;
+        [SerializeField]
+        private int _defaultPenalty = 5;
         [SerializeField]
         private float _slotMaxGrowthPercentage = 20;
         [SerializeField]
@@ -85,12 +89,16 @@ namespace MenuUI.Scripts.SoulHome
         private Room _roomInfo;
         private SoulHomeController _controller;
         private Transform _towerCamera;
+        private GridNode[,] _grid;
 
         private List<FurnitureSlot> _currentSlotValidity;
+        private List<Vector2Int> _walkableSlots = new();
 
         public Room RoomInfo { get => _roomInfo;}
         public int SlotRows { get => _slotRows;}
         public int SlotColumns { get => _slotColumns;}
+        public GridNode[,] Grid { get => _grid;}
+        public List<Vector2Int> WalkableSlots { get => _walkableSlots;}
         public SoulHomeController Controller { get => _controller;}
 
         void Start()
@@ -301,6 +309,149 @@ namespace MenuUI.Scripts.SoulHome
             }
 
             if (_roomInfo.Furnitures.Count > 0) InitialSetFurniture();
+
+            UpdateGrid();
+        }
+
+        public void UpdateGrid()
+        {
+            int columns = _slotColumns;
+            int rows = _slotRows;
+            if (_grid == null)
+            {
+                _grid = new GridNode[columns, rows];
+            }
+            
+            _walkableSlots.Clear();
+
+            for (int r = 0; r < rows; r++)
+            {
+                Transform rowTransform = transform.Find("FurniturePoints").Find("FloorFurniturePoints").GetChild(r);
+
+                for (int c = 0; c < columns; c++)
+                {
+                    FurnitureSlot slot = rowTransform.GetChild(c).GetComponent<FurnitureSlot>();
+
+                    //bool walkable = (slot.Furniture == null);
+                    //_grid[c, r] = new GridNode(new Vector2Int(c, r), walkable);
+
+                    if (_grid[c, r] == null)
+                    {
+                        _grid[c, r] = new GridNode(new Vector2Int(c, r), false);
+                    }
+                    GridNode node = _grid[c, r];
+                    
+                    if (slot.Furniture != null)
+                    {
+                        node.IsFurniture = true;
+                        node.Furniture = slot.Furniture;
+                    }
+                    _grid[c, r] = node;
+                }
+            }
+            int gridWidth = _grid.GetLength(0);
+            int gridHeight = _grid.GetLength(1);
+            CalculatePenalties();
+
+            // only add slots with no penalty to walkableslots
+            // need to make exception if there are none
+            for (int x = 0; x < gridWidth; x++)
+            {
+                for (int y = 0; y < gridHeight; y++)
+                {
+                    if (!_grid[x, y].IsFurniture && _grid[x, y].penalty == 0)
+                    {
+                        _walkableSlots.Add(new Vector2Int(x, y));
+                    }
+                }
+            }
+        }
+
+        // Add a penalty to tiles to the right or left of furniture, and double penalty if there is furniture to the left and the right,
+        // and also add a penalty to the bottom-right corner + 1 node to the left, and bottom-left corner and 1 node from that to the right, 
+        // so the avatar only goes there if there is no other way (makes clipping with furniture less likely, and avatar only chooses to "slip"
+        // through the corner of the furniture if there is absolutely no other way to the end position)
+        private void CalculatePenalties()
+        {
+            foreach (GridNode node in _grid)
+            {
+                node.penalty = 0;
+            }
+
+            int gridWidth = _grid.GetLength(0);
+            int gridHeight = _grid.GetLength(1);
+
+            for (int x = 0; x < gridWidth; x++)
+            {
+                for (int y = 0; y < gridHeight; y++)
+                {
+                    GridNode currentNode = _grid[x, y];
+                    // skip cells that are not furniture
+                    if (!currentNode.IsFurniture) continue;
+                    // Will be true for 2 pieces of the same furniture, which is not ideal
+                    bool isSameFurnitureAbove = (y > 0 &&
+                                                 _grid[x, y - 1].IsFurniture &&
+                                                 _grid[x, y - 1].Furniture == currentNode.Furniture);
+
+                    if (y == 0)
+                    {
+                        currentNode.penalty = 100;
+                    }
+                    else if (!_grid[x, y - 1].IsFurniture || !isSameFurnitureAbove)
+                    {
+                        currentNode.penalty = 150;
+                    }
+                    else
+                    {
+                        currentNode.penalty = 800;
+                    }
+
+                    // apply penalty to the left and right to the furniture
+                    ApplyPenalty(x - 1, y, _defaultPenalty);
+                    ApplyPenalty(x + 1, y, _defaultPenalty);
+
+
+                    // if there is no furniture below
+                    if (y < gridHeight - 1 && !_grid[x, y + 1].IsFurniture)
+                    {
+                        // if there is no furniture to the left (meaning this is the bottom-left corner)
+                        if (x > 0 && !_grid[x - 1, y].IsFurniture)
+                        {
+                            // The bottom-left corner
+                            ApplyPenalty(x - 1, y + 1, _defaultPenalty);
+                            // The cell above it
+                            ApplyPenalty(x - 1, y, _defaultPenalty);
+                            // The cell to the right of the bottom-left corner
+                            ApplyPenalty(x, y + 1, _defaultPenalty * 2);
+                        }
+                        // if there is no furniture to the right (meaning this is the bottom-right corner)
+                        if (x < gridWidth - 1 && !_grid[x + 1, y].IsFurniture)
+                        {
+                            // The bottom-right corner
+                            ApplyPenalty(x + 1, y + 1, _defaultPenalty);
+                            // The cell above it
+                            ApplyPenalty(x + 1, y, _defaultPenalty);
+                            // The cell to the left of the bottom-right corner
+                            ApplyPenalty(x, y + 1, _defaultPenalty * 2);
+                        }
+                    }
+                }
+            }
+        }
+
+        // applies penalty if inside grid bounds
+        public void ApplyPenalty(int x, int y, int amount)
+        {
+            int gridWidth = _grid.GetLength(0);
+            int gridHeight = _grid.GetLength(1);
+
+            if (x >= 0 && x < gridWidth && y >= 0 && y < gridHeight)
+            {
+                if (!_grid[x, y].IsFurniture)
+                {
+                    _grid[x, y].penalty += amount;
+                }
+            }
         }
 
         private void InitialSetFurniture()
